+++ js/data.js
@@ -1,0 +1,19 @@
+import { randInt } from "./util.js";
+
+const generatePhoto = (n) => ({
+  id: n,
+  url: `photos/${  n  }.jpg`,
+  description: `jpg № ${ n }`,
+  likes: randInt(15, 200),
+  comments: randInt(0, 200),
+})
+
+export const generatePhotos = (n) => {
+  const data = [];
+  for (let i = 0; i < n; i++) {
+    data[i] = photo(i+1);
+  }
+  return data;
+}
+
+photoData(25);
